requirements.txt
aiofiles==0.4.0
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-bootstrap3==11.0.0
django-crispy-forms==1.7.2
fasteners==0.15
idna==2.8
monotonic==1.5
multidict==4.5.2
node-semver==0.6.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
semantic-version==2.6.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
toolz==0.9.0
urllib3==1.25.3
virtualenv==16.6.0
yarl==1.3.0
zipfile36==0.1.3
